requirements.txt
# Core
python-dotenv==1.0.0
python-decouple==3.8

# AI & Content Generation
google-generativeai==0.3.0
requests==2.31.0

# Image/Video Processing
Pillow==10.1.0
opencv-python==4.8.1.78
ffmpeg-python==0.2.1

# Social Media APIs
instagrapi==2.0.1
facebook-sdk==3.0.2
tweepy==4.14.0
python-telegram-bot==20.3
praw==7.7.0  # Reddit

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Database (opcional)
sqlalchemy==2.0.23
sqlite3

# Scheduling
APScheduler==3.10.4
schedule==1.2.0

# Logging
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Development
black==23.12.0
flake8==6.1.0
pylint==3.0.3

